feat(checker.go): add column checking feature
The function takes in the file and check for the presence of columns
that are empty. If a colum is empty in all rows its index is added to
the return slice of the function.

test(checker_test.go): add multiple common test cases

The added test cases should be cover most of the expected tool usage

docs(checker.go): add GoDoc comment with explanation and example usage